outsourced functions from GameConnection to GameController
